+++ dcron_theme/base.html
@@ -57,7 +57,7 @@
 
         <blockquote>
           <div class="container">
-            <p>Dcron is a system service that runs scheduled jobs at given intervals or times, just like the cron unix service but distributed in several machines in a cluster. If a machine fails (the leader), a follower will take over and keep running the scheduled jobs without human intervention. Dcron is Open Source and freely available.</p>
+            <p>Dkron is a system service that runs scheduled jobs at given intervals or times, just like the cron unix service but distributed in several machines in a cluster. If a machine fails (the leader), a follower will take over and keep running the scheduled jobs without human intervention. Dcron is Open Source and freely available.</p>
           </div>
         </blockquote>
