+++ dkron/flag_slice_value.go
@@ -1,4 +1,4 @@
-package dcron
+package dkron
 
 import "strings"
