+++ dkron/flag_slice_value_test.go
@@ -1,4 +1,4 @@
-package dcron
+package dkron
 
 import (
 	"flag"
