+++ dkron/log.go
@@ -1,4 +1,4 @@
-package dcron
+package dkron
 
 import (
 	"github.com/Sirupsen/logrus"
